updated user model to support plant addition
