+++ client/src/components/Search/search.js
@@ -1,5 +1,5 @@
-import React, { useState, useEffect } from "react";
-import { Col, Row, Container } from '../../components/Grid';
+import React from "react";
+import { Row } from '../../components/Grid';
 import './style.css'
 
 function SearchResults(props) {
@@ -8,7 +8,7 @@ function SearchResults(props) {
     <Row>
       <div className="overflow-auto col-6 col-md-6">
       { props.results ? props.results.map(item => (
-        <div className="card card-body">
+        <div key={item.id} className="card card-body">
           <div className="welcome-container">
             <h4>{item.common_name}</h4>
               <p>{item.scientific_name}</p>
