+++ models/Plant.js
@@ -1,23 +1,23 @@
-const mongoose = require('mongoose');
+// const mongoose = require('mongoose');
 
-const PlantSchema = new mongoose.Schema({
-  name: {
-    type: String,
-    required: true
-  },
-  scientific_name: {
-    type: String
-  },
-  plant_id: {
-    type: Number,
-    required: true
-  },
-  date: {
-    type: Date,
-    default: Date.now
-  }
-});
+// const PlantSchema = new mongoose.Schema({
+//   name: {
+//     type: String,
+//     required: true
+//   },
+//   scientific_name: {
+//     type: String
+//   },
+//   plant_id: {
+//     type: Number,
+//     required: true
+//   },
+//   date: {
+//     type: Date,
+//     default: Date.now
+//   }
+// });
 
-const Plant = mongoose.model('Plant', PlantSchema);
+// const Plant = mongoose.model('Plant', PlantSchema);
 
-module.exports = Plant;
+// module.exports = Plant;
