+++ models/User.js
@@ -1,5 +1,27 @@
 const mongoose = require('mongoose');
 
+
+const PlantSchema = new mongoose.Schema({
+  name: {
+    type: String,
+    required: true
+  },
+  scientific_name: {
+    type: String
+  },
+  plant_id: {
+    type: Number,
+    required: true
+  },
+  date: {
+    type: Date,
+    default: Date.now
+  }
+});
+
+const Plant = mongoose.model('Plant', PlantSchema);
+
+
 const UserSchema = new mongoose.Schema({
   name: {
     type: String,
@@ -16,9 +38,12 @@ const UserSchema = new mongoose.Schema({
   date: {
     type: Date,
     default: Date.now
-  }
+  },
+  plants: [PlantSchema]
 });
 
 const User = mongoose.model('User', UserSchema);
 
-module.exports = User;
+
+
+module.exports = User, Plant;
