+++ views/dashboard.ejs
@@ -1,3 +1,13 @@
 <h1 class="mt-4">Dashboard</h1>
 <p class="lead mb-3">Welcome <%= user.name %></p>
 <a href="/users/logout" class="btn btn-secondary">Logout</a>
+
+<button href="/api" onclick="subs()">Test</button>
+
+<script>
+function subs(){
+    fetch("/api", {
+        method:"POST"
+        }).then(response=>{console.log(response)})
+        }
+</script>
